DataEngineeringProject3/requirements.txt
kafka-python==2.0.2
requests==2.27.1
json5==0.9.6
